Version final

+++ backend/controllers/tareasControllers.js
@@ -3,7 +3,7 @@ const Tarea = require('../models/tareasModel')
 
 const getTareas = asyncHandler(async (req, res) => {
 
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({ user: req.user.id })
 
     res.status(200).json(tareas)
 })
@@ -17,7 +17,8 @@ const createTareas = asyncHandler(async (req, res) => {
     }
 
     const tarea = await Tarea.create({
-        descripcion: req.body.descripcion
+        descripcion: req.body.descripcion,
+        user: req.user.id
     })
 
     res.status(201).json(tarea)
@@ -32,9 +33,14 @@ const updateTareas = asyncHandler(async (req, res) => {
         throw new Error('Esa tarea no existe')
     }
 
-    const tareaModificada = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
-
-    res.status(200).json(tareaModificada)
+    //Nos tenemos que asegurar que la tarea pertenece al usuario logueado
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Usuario no autorizado')
+    } else {
+        const tareaModificada = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
+        res.status(200).json(tareaModificada)
+    }
 })
 
 const deleteTareas = asyncHandler(async (req, res) => {
@@ -46,9 +52,14 @@ const deleteTareas = asyncHandler(async (req, res) => {
         throw new Error('Esa tarea no existe')
     }
 
-    await Tarea.deleteOne(tarea)
-
-    res.status(200).json({ id: req.params.id })
+    //Nos aseguramos que la tarea pertenezca al usuario logueado
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Usuario no autorizado')
+    } else {
+        await Tarea.deleteOne(tarea)
+        res.status(200).json({ id: req.params.id })
+    }
 })
 
 module.exports = {
